+++ test/test.py
@@ -27,10 +27,7 @@ def setup():
         import pygments  # noqa
     except ImportError:
         pygments_dir = join(top_dir, "deps", "pygments")
-        if sys.version_info[0] <= 2:
-            sys.path.insert(0, pygments_dir)
-        else:
-            sys.path.insert(0, pygments_dir + "3")
+        sys.path.insert(0, pygments_dir + "3")
 
 if __name__ == "__main__":
     logging.basicConfig()
@@ -42,7 +39,7 @@ if __name__ == "__main__":
         try:
             mod = importlib.import_module(extra_lib)
         except ImportError:
-            warnings.append("skipping %s tests ('%s' module not found)" % (extra_lib, extra_lib))
+            warnings.append("skipping {} tests ('{}' module not found)".format(extra_lib, extra_lib))
             default_tags.append("-%s" % extra_lib)
         else:
             if extra_lib == 'pygments':
@@ -51,7 +48,7 @@ if __name__ == "__main__":
                     tag = "pygments<2.14"
                 else:
                     tag = "pygments>=2.14"
-                warnings.append("skipping %s tests (pygments %s found)" % (tag, mod.__version__))
+                warnings.append("skipping {} tests (pygments {} found)".format(tag, mod.__version__))
                 default_tags.append("-%s" % tag)
 
     retval = testlib.harness(testdir_from_ns=testdir_from_ns,